fix: useSearchParams로 인한 에러 Suspense 이용해서 해결
- page 내에 바로 내용 넣지 않고 컨테이너 하나 따로 파서 넣도록 수정
